Create Google Scraper Basic
This is the golden 30 lines of my code. This here can be looped on multiple virtual computers and will scrape google trends. 

It is incredibly inefficient vs what you might find for other websites but google works very hard to stop web-scrapers and so this is the only thing that I've found works. This was made after consulting a couple top programmers.